datarun.py now does sbatch runs that don't write .csp or ALTI to file.
pyrescal.py was moved to datarun.py.
datarun.py was moved to rundirectory.py.
rescal_utilities can be made to update the real_data directory path.
rescal, in space.c was updated so that it doesn't try to
compress the .csp file if it never wrote it due to piping.
example_pyrescal is an executable python script that can
be called with a .sbatch file
